+++ server/logic/auth.js
@@ -2,7 +2,12 @@ function isLoggedIn(req, res, next) {
   if (req.isAuthenticated()) {
     next();
   } else {
-    next({ status: 403, message: "Unauthorized" });
+    console.log('hello');
+    /* next({ status: 403, message: 'Unauthorized' }); */
+    return res.status(403).json({
+      success: false,
+      message: 'You are not logged in',
+    });
   }
 }